s2: Haxx Goodix blobs to work on 10
Our Goodix blobs were compiled for Android 6.0.1, which wasn't issue in Android >=9, but now started to be one.
We often get stack corruption even though stack isn't really corrupted, sometimes mutex is locked with nullptr and a lot of stuff like that.

Everything was patched by hand in hex, using IDA as help.

Signed-off-by: Kuba Wojciechowski <nullbytepl@gmail.com>